Move file to api directory
