+++ app/src/main/java/me/jj97181818/dangongcha/api/GetRequest.java
@@ -1,4 +1,4 @@
-package me.jj97181818.dangongcha;
+package me.jj97181818.dangongcha.api;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
