+++ app/src/main/java/me/jj97181818/dangongcha/api/Infos.java
@@ -1,4 +1,4 @@
-package me.jj97181818.dangongcha;
+package me.jj97181818.dangongcha.api;
 
 import java.io.Serializable;
 import java.util.ArrayList;
